Fixed issues and polished implementation of mod install process
NOTE : As of now, file copying is used. In the future, mod installation could be handled through symlinks, which is what would make most sense, but that requires admin perms on windows, so that means I need to ask users to elevate the program's privileges for that to be possible. Maybe add it as a flag or something.