Created SimpleCopyFileHandler class
The SimpleCopyFileHandler class implements a simpler version of the CopyFileHandler class.

This appears to be faster despite the fact that it does not perform all of the de-duplication logic... fuck...

Well, faster only when it comes to installing large quantities of files. When dealing with multiple smaller mods that add a few files each, the old implementation is pretty fast. But that implementation misserably fails at properly dealing with larger mods, such as full game overhauls that override tons of files, probably due to the string comparisons and look ups being absurdly slow and linear.

With this new implementation, we have a lot of file overwriting, but at least it is surprisingly faster to install large overhaul mods with tons of assets.

Maybe this is the way we should work?

TODO : Consider multi-install setup rather than single install setup for further improvements in performance and greater support for custom executable mods.