Moved all code into Legacy directory
This change has been performed because a new idea for mcow mm has come up.

The new design will be closer in design to how other mod managers work.

The new philosophy for mcow mm is going to be specially inspired by popular mod managers / organizers for games like Skyrim, such as MO2, primarily because of the user friendly interface that the program offers and efficient resource usage.

Note that a VFS (virtual file system) implementation could be considered in the future, but as of now, the main code will make use of hardlinks and symlinks + executable copies.