Improved exception messages on mod manager constructor
The "does not exist" statement sounds too harsh and could be wrong in the event that the user or the program does not have the right permissions to access the directories.

The exception messages have been changed to say that it could not be found instead (maybe change to could not be accessed?) so that it is clear that if the path does exist, the program simply does not have the ability to access it.