Added function to mod manager to delete manifest
This function will be used when re-generating a profile's files.
This will allow external programs such as mpupgui to monitor the existence of this file for updates, allowing it to know whether a profile install has finished or not.

If a profile install is abruptly terminated and the profile was not properly installed, the manifest will not be created at the end of the operation, which will allow other programs to detect the fact that the profile never finished installing.