Modified cmdvar_ProfilesToApply to be cleared at the end of the execution
Allows the command handler to be reused in more complex programs.

This would probably be more cleanly implemented by having some generic reset logic that is applied each time the parse command is called from the parser so that multiple handlers and parsers could be used, but this is ok for now.